bug fix: float conversion to int

+++ cal.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 
-int sum(int a, int b) {
+float sum(float a, float b) {
     return (a+b);
 }
